Select region based on luminance instead of single channel values

+++ source/blender/makesdna/DNA_node_types.h
@@ -879,7 +879,7 @@ typedef struct NodeBilateralBlurData {
 typedef struct NodeKuwaharaData {
   short kernel_size;
   short variation;
-  float sigma;
+  int sigma;
 } NodeKuwaharaData;
 
 typedef struct NodeAntiAliasingData {
